requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf==4.25.3
faster-whisper>=1.0.0
soundfile
sentence-transformers
